+++ libs/JSystem/JAudio2/JASWaveArcLoader.cpp
@@ -6,6 +6,7 @@
 #include "JSystem/JAudio2/JASWaveArcLoader.h"
 #include "JSystem/JAudio2/JASDvdThread.h"
 #include "JSystem/JAudio2/JASTaskThread.h"
+#include "JSystem/JAudio2/JASMutex.h"
 #include "JSystem/JKernel/JKRDvdAramRipper.h"
 #include "MSL_C/string.h"
 #include "dolphin/dvd/dvd.h"
